Avoid mixing variables and indices. Handle unbounded cases.
Variable names (bounded by m+n) were used directly as indices in
vectors/matrices of dimension m. They are now translated back first by
computing the inverse mapping of that stored in bs/nbs. To make the
distinction clearer variables are now wrapped in a struct.

The procedure for choosing a leaving variable is modified slightly to
guarantee that -1 is returned when all choices are inf. This avoids
infinite cycling when the solution is unbounded.